# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Loja Online",
        "description": "Documentação da API para gerenciamento de produtos.",
        "contact": {
            "email": "seu@email.com"
        },
        "version": "1.0"
    },
    "paths": {
        "/api/produtos": {
            "get": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Lista todos os produtos",
                "operationId": "10841ce9f899ba906a6a4f19fd0a5fbe",
                "parameters": [
                    {
                        "name": "categoria",
                        "in": "query",
                        "description": "Filtrar por categoria",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de produtos"
                    }
                }
            },
            "post": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Cria um novo produto",
                "operationId": "15f14d72ddfbf1fea0baab03d8a81aa0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "descricao",
                                    "preco",
                                    "quantidade",
                                    "categoria"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string"
                                    },
                                    "descricao": {
                                        "type": "string"
                                    },
                                    "preco": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "quantidade": {
                                        "type": "integer"
                                    },
                                    "categoria": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Produto criado"
                    }
                }
            }
        },
        "/api/produtos/{id}": {
            "get": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Exibe um produto específico",
                "operationId": "93b903542291f0b00d6cbf08269b86c8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados do produto"
                    },
                    "404": {
                        "description": "Produto não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Atualiza um produto",
                "operationId": "0f8bfb14fd7ee25ea71a034a7ad2fc6c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "type": "string"
                                    },
                                    "descricao": {
                                        "type": "string"
                                    },
                                    "preco": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "quantidade": {
                                        "type": "integer"
                                    },
                                    "categoria": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Produto atualizado"
                    },
                    "404": {
                        "description": "Produto não encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Exclui um produto",
                "operationId": "ff7c08fe1f3dee747fe4b668755be73a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Produto excluído"
                    },
                    "404": {
                        "description": "Produto não encontrado"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Produtos",
            "description": "Produtos"
        }
    ]
}